Интеграция Arduino FS обобщена. Потокозащищенная.
